Split transport epics
